+++ src/NewPicker/PickerInput/SinglePicker.tsx
@@ -3,7 +3,7 @@ import useLayoutEffect from 'rc-util/lib/hooks/useLayoutEffect';
 import omit from 'rc-util/lib/omit';
 import pickAttrs from 'rc-util/lib/pickAttrs';
 import * as React from 'react';
-import useToggleDates from '../hooks/useToggleDates';
+import useToggleDates from '../../hooks/useToggleDates';
 import type {
   BaseInfo,
   InternalMode,
