+++ src/hooks/useTimeInfo.ts
@@ -1,9 +1,9 @@
 import { warning } from 'rc-util';
 import * as React from 'react';
-import type { GenerateConfig } from '../../generate';
-import { leftPad } from '../../utils/miscUtil';
-import type { DisabledTimes, SharedTimeProps } from '../../interface';
-import { findValidateTime } from '../PickerPanel/TimePanel/TimePanelBody/util';
+import type { GenerateConfig } from '../generate';
+import { leftPad } from '../utils/miscUtil';
+import type { DisabledTimes, SharedTimeProps } from '../interface';
+import { findValidateTime } from '../NewPicker/PickerPanel/TimePanel/TimePanelBody/util';
 
 export type Unit<ValueType = number | string> = {
   label: React.ReactText;
